Keyboard Controls + layout Changes + visuals
Press left and right arrow keys to choose an answer, and space bar to navigate to the next question.

There is still a bug for when you change from keyboard controls to clicking the buttons -  an incorrect choice still invokes "correct answer" on screen. Looking for a solution for this.